// client/src/features/account/register/register.component.ts
import { Component, inject, output, signal } from '@angular/core';
import { AbstractControl, FormBuilder, FormControl, FormGroup, ReactiveFormsModule, ValidationErrors, ValidatorFn, Validators } from '@angular/forms';
import { RegisterCreds, User } from '../../../types/user';
import { AccountServiceService } from '../../../core/services/account-service.service';
import { TextInputComponent } from "../../../shared/text-input/text-input.component";
import { Router } from '@angular/router';

@Component({
  selector: 'app-register',
  imports: [ReactiveFormsModule, TextInputComponent],
  templateUrl: './register.component.html',
  styleUrl: './register.component.css'
})
export class RegisterComponent{
  
  private accountService = inject(AccountServiceService);
  private router  = inject(Router);
  private fb = inject(FormBuilder);
  cancelRegister = output<boolean>();
  protected creds = {} as RegisterCreds;
  protected credentialsForm: FormGroup;
  protected profileForm: FormGroup;
  protected currentStep = signal(1);
  protected validationErrors = signal<string[]>([]);

  constructor(){
      this.credentialsForm = this.fb.group({
      email: ['', [Validators.required,Validators.email]],
      displayName: ['', Validators.required],
      password: ['',[Validators.required,
        Validators.minLength(4), Validators.maxLength(8)]],
      confirmpassword:['',[Validators.required,this.matchValues('password')]]
    });

    this.profileForm = this.fb.group({
      gender: ['male',Validators.required],
      dateOfBirth: ['',Validators.required],
      city: ['',Validators.required],
      country: ['',Validators.required],

    })

    this.credentialsForm.controls['password'].valueChanges.subscribe(() => {
      this.credentialsForm.controls['confirmPassword'].updateValueAndValidity();
    })
  }

  matchValues(matchTo: string): ValidatorFn{
    return (control: AbstractControl): ValidationErrors | null => {
      const parent = control.parent;
      if(!parent) return null;

      const matchValue = parent.get(matchTo)?.value;
      return control.value === matchValue ? null : {passwordMismatch: true}
    }
  }

  nextStep(){
    if(this.credentialsForm.valid){
      this.currentStep.update(prevStep => prevStep+1);
    }
  }

  prevStep() {
    this.currentStep.update(prevStep => prevStep - 1);
  }

  getMaxDate(){
    const today = new Date();
    today.setFullYear(today.getFullYear() - 18);
    return today.toISOString().split('T')[0];
  }

  register(){
    if(this.profileForm.valid && this.credentialsForm.valid){
      const formData = {...this.credentialsForm.value, ...this.profileForm.value};
      // console.log('Form data: ',formData)
      this.accountService.register(formData).subscribe({
      next: () => {
        this.router.navigateByUrl('/members');
      },
      error: error => {
        console.error(error);
        this.validationErrors.set(error);
      }
      
     })
    }
 
  }

  cancel(){
    this.cancelRegister.emit(false);
    console.log('cancelled')
  }
}
